requirements.txt
psycopg2-binary==2.9.9
PyYAML==6.0.2
pandas==2.2.2
numpy==1.26.4
tqdm==4.66.5
python-dotenv==1.0.1
pytest==8.3.3
pytest-cov==5.0.0
